SM Planting/Voting, TreeWindow fixes
Added planting and voting for social media logins.
getCurrentUserObj().done(function(userObj) { ... Queries the database
based on current login method, returning the user to the callback
function as userObj to allow the same code to be used in both cases.
Also several fixes for the new TreeWindows
-- fixed the vote button firing off multiple functions because it's not
a unique ID each time the window was opened another function was
attached to the button used by every window
-- fixed votes not updating when reopening a window, tree only referred
to a local copy, not the real tree in the database
-- fixed tree window being used even if the tree hasn't been confirmed
yet
Also added tree names to the current info window planter for use until
new tree planter is pushed.
Still need to do mobile stuff